week6/pyproject.toml
[tool.poetry]
name = "week6"
version = "0.1.0"
description = ""
authors = ["AliceKuo <ming-chieh.kuo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
itsdangerous = "^2.2.0"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
